+++ lib/src/options.dart
@@ -1,11 +1,19 @@
 import 'dart:ui';
 
+import 'package:collection/collection.dart';
 import 'package:flutter/foundation.dart';
 
-typedef ArgumentFormatter = String Function(
-  Object value,
-  String? format,
+// ignore_for_file: lines_longer_than_80_chars
+
+/// The formatter signature for [I18NextOptions.formats].
+///
+/// The [value] can be null if the variable wasn't evaluated properly, giving a
+/// chance for the formatter to do something
+typedef ValueFormatter = Object? Function(
+  Object? value,
+  Map<String, Object> formatOptions,
   Locale locale,
+  I18NextOptions options,
 );
 
 typedef MissingKeyHandler = String? Function(
@@ -34,12 +42,15 @@ class I18NextOptions with Diagnosticable {
     this.keySeparator,
     this.interpolationPrefix,
     this.interpolationSuffix,
-    this.interpolationSeparator,
+    this.formatSeparator,
+    this.formatterValues,
+    this.formats,
+    this.optionsSeparator,
+    this.optionValueSeparator,
     this.nestingPrefix,
     this.nestingSuffix,
     this.nestingSeparator,
     this.pluralSuffix,
-    this.formatter,
     this.missingKeyHandler,
     this.translationFailedHandler,
   }) : super();
@@ -52,12 +63,18 @@ class I18NextOptions with Diagnosticable {
     keySeparator: '.',
     interpolationPrefix: '{{',
     interpolationSuffix: '}}',
-    interpolationSeparator: ',',
+    formatSeparator: ',',
+    formatterValues: {
+      'true': true,
+      'false': false,
+    },
+    formats: {},
+    optionsSeparator: ';',
+    optionValueSeparator: ':',
     nestingPrefix: r'$t(',
     nestingSuffix: ')',
     nestingSeparator: ',',
     pluralSuffix: 'plural',
-    formatter: null,
     missingKeyHandler: null,
     translationFailedHandler: null,
   );
@@ -110,17 +127,40 @@ class I18NextOptions with Diagnosticable {
   /// By default they are '{{' and '}}' respectively and can't be null but
   /// can be empty.
   ///
-  /// [interpolationSeparator] is used to separate the variable's
+  /// [formatSeparator] is used to separate the variable's
   /// name from the format (if any). Defaults to ',' and cannot be null nor
   /// empty (otherwise it'll match every char in the interpolation).
   ///
-  /// ```
-  /// - '{{title}}' name = 'title, format = null
-  /// - '{{title, uppercase}}' name = 'title', format = 'uppercase'
-  /// ```
-  final String? interpolationPrefix,
-      interpolationSuffix,
-      interpolationSeparator;
+  /// - '{{title}}'
+  /// - '{{some.variable.name}}'
+  /// - '{{some.variable.name, format1, format2}}'
+  /// - '{{some.variable.name, format(option: value)}}'
+  final String? interpolationPrefix, interpolationSuffix, formatSeparator;
+
+  /// [formats] is called when an interpolation has been found and it was marked
+  /// with formatting options.
+  ///
+  /// If the format doesn't exist then it either moves to the next format or
+  /// the value is returned as is (in string form: [Object.toString]).
+  final Map<String, ValueFormatter>? formats;
+
+  /// Helper that maps known or common [formats] values to a specific value.
+  ///
+  /// e.g.:
+  /// `{{formatName(option: true)}}`
+  /// Maps the "true" string to a bool value while parsing.
+  final Map<String, Object>? formatterValues;
+
+  /// Options are delimited for the [formats], and are defined
+  /// between '(' and ')'.
+  ///
+  /// [optionsSeparator] separates all the available options between the
+  /// prefix and suffix, while the [optionValueSeparator] separates the key
+  /// values for each option.
+  ///
+  /// "Some format {{value, formatName(optionName: optionValue)}}",
+  /// "Some format {{value, formatName(option1Name: option1Value; option2Name: option2Value)}}"
+  final String? optionsSeparator, optionValueSeparator;
 
   /// [nestingPrefix] and [nestingSuffix] are the deliminators for nesting
   /// mechanism. By default they are '$t(' and ')' respectively and can't be
@@ -138,13 +178,6 @@ class I18NextOptions with Diagnosticable {
   /// i18Next.t('key1') // "Hello World!"
   /// ```
   final String? nestingPrefix, nestingSuffix, nestingSeparator;
-
-  /// [formatter] is called when an interpolation has been found and is ready
-  /// for substitution.
-  ///
-  /// Defaults to [defaultFormatter], which simply returns the value itself in
-  /// String form ([Object.toString]).
-  final ArgumentFormatter? formatter;
 
   /// The default behavior is to just return the key itself that was used in
   /// [I18Next.t].
@@ -171,12 +204,18 @@ class I18NextOptions with Diagnosticable {
       pluralSuffix: other.pluralSuffix ?? pluralSuffix,
       interpolationPrefix: other.interpolationPrefix ?? interpolationPrefix,
       interpolationSuffix: other.interpolationSuffix ?? interpolationSuffix,
-      interpolationSeparator:
-          other.interpolationSeparator ?? interpolationSeparator,
+      formatSeparator: other.formatSeparator ?? formatSeparator,
+      formatterValues: other.formatterValues ?? formatterValues,
+      formats: formats == null
+          ? other.formats
+          : other.formats == null
+              ? formats
+              : {...?formats, ...?other.formats},
+      optionsSeparator: other.optionsSeparator ?? optionsSeparator,
+      optionValueSeparator: other.optionValueSeparator ?? optionValueSeparator,
       nestingPrefix: other.nestingPrefix ?? nestingPrefix,
       nestingSuffix: other.nestingSuffix ?? nestingSuffix,
       nestingSeparator: other.nestingSeparator ?? nestingSeparator,
-      formatter: other.formatter ?? formatter,
       missingKeyHandler: other.missingKeyHandler ?? missingKeyHandler,
       translationFailedHandler:
           other.translationFailedHandler ?? translationFailedHandler,
@@ -194,11 +233,14 @@ class I18NextOptions with Diagnosticable {
     String? pluralSuffix,
     String? interpolationPrefix,
     String? interpolationSuffix,
-    String? interpolationSeparator,
+    String? formatSeparator,
+    Map<String, Object>? formatterValues,
+    Map<String, ValueFormatter>? formats,
+    String? optionsSeparator,
+    String? optionValueSeparator,
     String? nestingPrefix,
     String? nestingSuffix,
     String? nestingSeparator,
-    ArgumentFormatter? formatter,
     MissingKeyHandler? missingKeyHandler,
     TranslationFailedHandler? translationFailedHandler,
   }) {
@@ -211,12 +253,14 @@ class I18NextOptions with Diagnosticable {
       pluralSuffix: pluralSuffix ?? this.pluralSuffix,
       interpolationPrefix: interpolationPrefix ?? this.interpolationPrefix,
       interpolationSuffix: interpolationSuffix ?? this.interpolationSuffix,
-      interpolationSeparator:
-          interpolationSeparator ?? this.interpolationSeparator,
+      formatSeparator: formatSeparator ?? this.formatSeparator,
+      formatterValues: formatterValues ?? this.formatterValues,
+      formats: formats ?? this.formats,
+      optionsSeparator: optionsSeparator ?? this.optionsSeparator,
+      optionValueSeparator: optionValueSeparator ?? this.optionValueSeparator,
       nestingPrefix: nestingPrefix ?? this.nestingPrefix,
       nestingSuffix: nestingSuffix ?? this.nestingSuffix,
       nestingSeparator: nestingSeparator ?? this.nestingSeparator,
-      formatter: formatter ?? this.formatter,
       missingKeyHandler: missingKeyHandler ?? this.missingKeyHandler,
       translationFailedHandler:
           translationFailedHandler ?? this.translationFailedHandler,
@@ -231,35 +275,43 @@ class I18NextOptions with Diagnosticable {
         keySeparator,
         interpolationPrefix,
         interpolationSuffix,
-        interpolationSeparator,
+        formatSeparator,
+        formatterValues,
+        formats,
+        optionsSeparator,
+        optionValueSeparator,
         nestingPrefix,
         nestingSuffix,
         nestingSeparator,
         pluralSuffix,
-        formatter,
         missingKeyHandler,
         translationFailedHandler,
       );
 
   @override
-  bool operator ==(Object other) =>
-      other.runtimeType == runtimeType &&
-      other is I18NextOptions &&
-      other.fallbackNamespaces == fallbackNamespaces &&
-      other.namespaceSeparator == namespaceSeparator &&
-      other.contextSeparator == contextSeparator &&
-      other.pluralSeparator == pluralSeparator &&
-      other.keySeparator == keySeparator &&
-      other.interpolationPrefix == interpolationPrefix &&
-      other.interpolationSuffix == interpolationSuffix &&
-      other.interpolationSeparator == interpolationSeparator &&
-      other.nestingPrefix == nestingPrefix &&
-      other.nestingSuffix == nestingSuffix &&
-      other.nestingSeparator == nestingSeparator &&
-      other.pluralSuffix == pluralSuffix &&
-      other.formatter == formatter &&
-      other.missingKeyHandler == missingKeyHandler &&
-      other.translationFailedHandler == translationFailedHandler;
+  bool operator ==(Object other) {
+    if (identical(this, other)) return true;
+    return other.runtimeType == runtimeType &&
+        other is I18NextOptions &&
+        other.fallbackNamespaces == fallbackNamespaces &&
+        other.namespaceSeparator == namespaceSeparator &&
+        other.contextSeparator == contextSeparator &&
+        other.pluralSeparator == pluralSeparator &&
+        other.keySeparator == keySeparator &&
+        other.interpolationPrefix == interpolationPrefix &&
+        other.interpolationSuffix == interpolationSuffix &&
+        other.formatSeparator == formatSeparator &&
+        other.formatterValues == formatterValues &&
+        const MapEquality().equals(other.formats, formats) &&
+        other.optionsSeparator == optionsSeparator &&
+        other.optionValueSeparator == optionValueSeparator &&
+        other.nestingPrefix == nestingPrefix &&
+        other.nestingSuffix == nestingSuffix &&
+        other.nestingSeparator == nestingSeparator &&
+        other.pluralSuffix == pluralSuffix &&
+        other.missingKeyHandler == missingKeyHandler &&
+        other.translationFailedHandler == translationFailedHandler;
+  }
 
   @override
   void debugFillProperties(DiagnosticPropertiesBuilder properties) {
@@ -272,12 +324,16 @@ class I18NextOptions with Diagnosticable {
       ..add(StringProperty('keySeparator', keySeparator))
       ..add(StringProperty('interpolationPrefix', interpolationPrefix))
       ..add(StringProperty('interpolationSuffix', interpolationSuffix))
-      ..add(StringProperty('interpolationSeparator', interpolationSeparator))
+      ..add(StringProperty('formatSeparator', formatSeparator))
+      ..add(StringProperty('formatterValues', formatterValues?.toString()))
+      ..add(StringProperty('formatter', formats?.toString()))
+      ..add(StringProperty('optionsSeparator', optionsSeparator))
+      ..add(StringProperty('optionValueSeparator', optionValueSeparator))
       ..add(StringProperty('nestingPrefix', nestingPrefix))
       ..add(StringProperty('nestingSuffix', nestingSuffix))
       ..add(StringProperty('nestingSeparator', nestingSeparator))
       ..add(StringProperty('pluralSuffix', pluralSuffix))
-      ..add(StringProperty('formatter', formatter?.toString()))
+      ..add(StringProperty('formatter', formats?.toString()))
       ..add(StringProperty('missingKeyHandler', missingKeyHandler?.toString()))
       ..add(StringProperty(
           'translationFailedHandler', translationFailedHandler?.toString()));