+++ CHANGELOG.md
@@ -10,6 +10,31 @@
 * Moves `I18NextOptions.defaultFormatter` as an internal `interpolator.dart` method.
 * Adds `MissingKeyHandler` and `TranslationFailedHandler` on `I18NextOptions` to allow custom handling if needed
   (default behavior is to return the key itself).
+* Refactors formatters **BREAKING CHANGE**:
+  * Now the formatters are registered by name rather than just by a single function:
+    ```dart
+    I18NextOptions(formats: {
+      'uppercase': (variable, variableOptions, locale, options) => variable?.toUpperCase(),
+      'lowercase': (variable, variableOptions, locale, options) => variable?.toUpperCase(),
+      // (...)
+    });
+    ```
+    The callback of the formatter may receive and return a null value so the next formats may have a chance to handle it
+    before returning to the interpolation call. Which if it is still null, will be considered a translation error, resulting 
+    in the original key being returned.
+    ```
+    "key": "Hello {{name, uppercase}}!" // "Hello WORLD!"
+    
+    // multiple formatters in sequence
+    "key": "Hello {{name, fmt1, fmt2, fmt3}}!" 
+    ```
+    And they also accept options if the format needs them:
+    ```json
+    {
+      "key": "Some format {{value, formatName}}",
+      "keyWithOptions": "Some format {{value, formatName(option1Name: option1Value; option2Name: option2Value)}}"
+    }
+    ```
 
 ## [0.5.2]